server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.0
redis==5.2.0
pydantic==2.10.0
pydantic-settings==2.7.0
openai==1.60.0
httpx==0.28.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
sse-starlette==2.1.0
tenacity==9.0.0
pyyaml==6.0.2
structlog==24.4.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.28.0
